tweaks

+++ src/main/java/rafer/Main.java
@@ -215,8 +215,13 @@ public class Main {
             path = dest.getRelative(destRoot);
             FileNode src = srcRoot.join(path);
             if (!src.exists()) {
-                console.info.println("D " + destRoot.getName() + "/" + path);
-                dest.deleteFile();
+                if (src.getParent().exists()) {
+                    console.info.println("D " + destRoot.getName() + "/" + path);
+                    dest.deleteFile();
+                } else {
+                    // e.g. if src only contains the current year, but the backups hold all the years
+                    console.info.println("not synced: " + src.getParent());
+                }
             }
         }
     }
